requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Flask-Cors==3.0.10
requests==2.26.0
python-dotenv==0.19.0
gunicorn==20.1.0
Pillow==8.3.1
numpy==1.21.1
google-api-python-client==2.19.1
google-cloud-vision==2.7.0
google-cloud-translate==3.7.1
